Refactor icons
