+++ tools/Dashboard/DevHome.Dashboard/Helpers/WidgetIconCache.cs
@@ -1,0 +1,119 @@
+﻿// Copyright (c) Microsoft Corporation and Contributors
+// Licensed under the MIT license.
+
+using System;
+using System.Collections.Generic;
+using System.Threading.Tasks;
+using CommunityToolkit.WinUI;
+using Microsoft.UI.Dispatching;
+using Microsoft.UI.Xaml;
+using Microsoft.UI.Xaml.Media;
+using Microsoft.UI.Xaml.Media.Imaging;
+using Microsoft.Windows.Widgets.Hosts;
+using Windows.Storage.Streams;
+
+namespace DevHome.Dashboard.Helpers;
+internal class WidgetIconCache
+{
+    private static DispatcherQueue _dispatcher;
+
+    private static Dictionary<string, BitmapImage> _widgetLightIconCache;
+    private static Dictionary<string, BitmapImage> _widgetDarkIconCache;
+
+    public WidgetIconCache(DispatcherQueue dispatcher)
+    {
+        _dispatcher = dispatcher;
+        _widgetLightIconCache = new Dictionary<string, BitmapImage>();
+        _widgetDarkIconCache = new Dictionary<string, BitmapImage>();
+    }
+
+    public async Task CacheAllWidgetIcons(WidgetCatalog widgetCatalog)
+    {
+        var widgetDefs = widgetCatalog.GetWidgetDefinitions();
+        foreach (var widgetDef in widgetDefs ?? Array.Empty<WidgetDefinition>())
+        {
+            await AddIconsToCache(widgetDef);
+        }
+    }
+
+    public async Task AddIconsToCache(WidgetDefinition widgetDef)
+    {
+        // Only cache icons for providers that we're including.
+        if (WidgetHelpers.IsIncludedWidgetProvider(widgetDef.ProviderDefinition))
+        {
+            var widgetDefId = widgetDef.Id;
+            try
+            {
+                Log.Logger()?.ReportDebug("DashboardView", $"Cache widget icon for {widgetDefId}");
+                var itemLightImage = await WidgetIconToBitmapImage(widgetDef.GetThemeResource(WidgetTheme.Light).Icon);
+                var itemDarkImage = await WidgetIconToBitmapImage(widgetDef.GetThemeResource(WidgetTheme.Dark).Icon);
+
+                // There is a widget bug where Definition update events are being raised as added events.
+                // If we already have an icon for this key, just remove and add again.
+                if (_widgetLightIconCache.ContainsKey(widgetDefId))
+                {
+                    _widgetLightIconCache.Remove(widgetDefId);
+                }
+
+                if (_widgetDarkIconCache.ContainsKey(widgetDefId))
+                {
+                    _widgetDarkIconCache.Remove(widgetDefId);
+                }
+
+                _widgetLightIconCache.Add(widgetDefId, itemLightImage);
+                _widgetDarkIconCache.Add(widgetDefId, itemDarkImage);
+            }
+            catch (Exception ex)
+            {
+                Log.Logger()?.ReportError("DashboardView", $"Exception in CacheWidgetIcons:", ex);
+                _widgetLightIconCache.Add(widgetDefId, null);
+                _widgetDarkIconCache.Add(widgetDefId, null);
+            }
+        }
+    }
+
+    public static BitmapImage GetWidgetIconForTheme(WidgetDefinition widgetDefinition, ElementTheme theme)
+    {
+        BitmapImage image;
+        if (theme == ElementTheme.Light)
+        {
+            _widgetLightIconCache.TryGetValue(widgetDefinition.Id, out image);
+        }
+        else
+        {
+            _widgetDarkIconCache.TryGetValue(widgetDefinition.Id, out image);
+        }
+
+        return image;
+    }
+
+    public static Brush GetBrushForWidgetIcon(WidgetDefinition widgetDefinition, ElementTheme theme)
+    {
+        var image = GetWidgetIconForTheme(widgetDefinition, theme);
+
+        var brush = new ImageBrush
+        {
+            ImageSource = image,
+        };
+        return brush;
+    }
+
+    public void RemoveIconsFromCache(string definitionId)
+    {
+        _widgetLightIconCache.Remove(definitionId);
+        _widgetDarkIconCache.Remove(definitionId);
+    }
+
+    private static async Task<BitmapImage> WidgetIconToBitmapImage(IRandomAccessStreamReference iconStreamRef)
+    {
+        var itemImage = await _dispatcher.EnqueueAsync(async () =>
+        {
+            using var bitmapStream = await iconStreamRef.OpenReadAsync();
+            var itemImage = new BitmapImage();
+            await itemImage.SetSourceAsync(bitmapStream);
+            return itemImage;
+        });
+
+        return itemImage;
+    }
+}
